requirements.txt
internetarchive>=5.0.0
requests>=2.0.0
streamlit>=1.30.0
rapidfuzz>=3.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
